+++ example/p5.d.ts
@@ -1,4 +1,3 @@
-// // This file will add both p5 instanced and global intellisence
 import * as p5Global from 'p5/global'
 import module = require('p5')
 export = module
