+++ package.json
@@ -6,8 +6,8 @@
     "types": "dist/index.d.ts",
     "scripts": {
         "clean": "rm dist/bundle.js",
-        "start": "snowpack dev --config snowpack.config.dev.js",
-        "build": "snowpack build --config snowpack.config.prod.js && tsc -p declarations.tsconfig.json && cp src/p5.d.ts dist/p5.d.ts"
+        "start": "snowpack dev --config dev.snowpack.config.js",
+        "build": "snowpack build --config prod.snowpack.config.js && tsc -p declarations.tsconfig.json && cp src/p5.d.ts dist/p5.d.ts"
     },
     "keywords": [
         "typescript",
